requirements.txt
dash
dash_core_components
dash_html_components
#io
datetime
#os
pandas
#dash_tables
plotly
plotly.express
mysql.connector
#socket
paramiko
#base64
cryptography
dash_extensions
#mysql-connector-python


